moved neuroscope sorting metadata to the more appropriate cellexplorer interface
